Move the app tenant resolving logic to a separate valve
